Updated Render + Refactored
Render - I was manually doing clipping, but I found out SDL can do that
for us, so I changed the render algorithm to take in account for this.
The code is much tighter, and the speed is the same as before.

Refactor - Moved the map class into it's own header where it should be.