Updated - Entity Events
MultipleChoice has been broken up into MultipleChoice(MC) and
MultipleChoiceChunk(MCC). MC  is used by an agent to setup a multiple
choice conversation, and then sends a MCC to another agent, after the
agent answers the question in the MCC's dialog, the MCC automatically
informs the sender.

A bit more work needs to be done before this is complete, but it
compiles and is almost there. Soon agents will easily be able to other
agents questions and get a response. Fun times.